pyproject.toml
[project]
name = "soccernet-reid"
version = "0.1.0"
description = "SoccerNet player re-identification — master's thesis experiments"
requires-python = ">=3.11,<3.13"
dependencies = [
    "pandas>=2.2",
    "pyarrow>=17.0",
    "pillow>=10.4",
    "numpy>=1.26,<2.3",
    "tqdm>=4.66",
    "SoccerNet>=0.1.60",
]

[project.optional-dependencies]
training = [
    "torch>=2.5",
    "torchvision>=0.20",
    "timm>=1.0.11",
    "pytorch-metric-learning>=2.7",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "wandb>=0.18",
    "albumentations>=1.4",
]
dev = [
    "pytest>=8.3",
    "ruff>=0.7",
    "ipython>=8.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/soccernet_reid"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
